dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
appdirs==1.4.4
    # via virtualenv
aspy.yaml==1.3.0
    # via pre-commit
atomicwrites==1.4.0
    # via pytest
attrs==19.3.0
    # via pytest
certifi==2020.12.5
    # via
    #   -c requirements.txt
    #   elasticsearch
cfgv==2.0.1
    # via pre-commit
coverage==5.2
    # via pytest-cov
distlib==0.3.1
    # via virtualenv
elasticsearch-dsl==7.4.0
    # via -r dev-requirements.in
elasticsearch==7.12.0
    # via
    #   -c requirements.txt
    #   elasticsearch-dsl
filelock==3.0.12
    # via virtualenv
identify==1.4.24
    # via pre-commit
importlib-metadata==1.7.0
    # via
    #   pluggy
    #   pre-commit
    #   pytest
    #   virtualenv
importlib-resources==3.2.1
    # via
    #   pre-commit
    #   virtualenv
more-itertools==5.0.0
    # via pytest
nodeenv==1.4.0
    # via pre-commit
packaging==20.4
    # via pytest
pathlib2==2.3.7.post1
    # via pytest
pluggy==0.13.1
    # via pytest
pre-commit==1.21.0
    # via -r dev-requirements.in
py==1.9.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest-cov==2.10.0
    # via -r dev-requirements.in
pytest==4.6.11
    # via
    #   -r dev-requirements.in
    #   pytest-cov
python-dateutil==2.8.2
    # via elasticsearch-dsl
pyyaml==5.3.1
    # via
    #   -c requirements.txt
    #   aspy.yaml
    #   pre-commit
six==1.15.0
    # via
    #   cfgv
    #   elasticsearch-dsl
    #   more-itertools
    #   packaging
    #   pathlib2
    #   pre-commit
    #   pytest
    #   python-dateutil
    #   virtualenv
toml==0.10.1
    # via pre-commit
urllib3==1.26.4
    # via
    #   -c requirements.txt
    #   elasticsearch
virtualenv==20.0.27
    # via pre-commit
wcwidth==0.2.5
    # via pytest
zipp==1.2.0
    # via
    #   importlib-metadata
    #   importlib-resources
